reorganize CLI scripts
