+++ .gitignore
@@ -1,3 +1,3 @@
-.idea/*
-docs/script/.env
-node_modules/*
+node_modules
+scripts/output/*.md
+.env
