--- docs/script/package.json
@@ -1,6 +1,0 @@
-{
-  "version": "1.0.0",
-  "scripts": {
-    "start": "node index.js"
-  }
-}
