+++ package.json
@@ -1,11 +1,19 @@
 {
   "version": "1.0.0",
   "scripts": {
-    "re-order": "npm run start --prefix docs/script"
+    "update-md-tables-in-doc": "ts-node ./scripts/update-md-tables-in-doc.ts",
+    "build-md-tables-from-openapi": "ts-node ./scripts/build-md-tables-from-openapi.ts",
+    "readme-fix-docs-order": "ts-node ./scripts/readme-fix-docs-order.ts"
   },
   "dependencies": {
     "dotenv": "^16.3.1",
+    "markdown-it": "^13.0.2",
     "minimist": "^1.2.8",
-    "node-fetch": "^2.7.0"
+    "node-fetch": "^2.7.0",
+    "ts-node": "^10.9.1",
+    "yup": "^1.1.1"
+  },
+  "devDependencies": {
+    "@types/node": "^20.7.0"
   }
 }
