+++ scripts/globals.t.ts
