--- tableWithNestedSchemaGenerator/README.md
@@ -1,5 +1,0 @@
-Example script to generate markdown tables describing selected objects that use "$ref" from OpenAPI file.
-
-Cli :
-- `npm i` - install required dependencies
-- `npm run stat` - generate markdown files into `output` folder
