--- tableWithNestedSchemaGenerator/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "openapi-nested-object-table",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.ts",
-  "scripts": {
-    "start": "ts-node ./index.ts ",
-    "updateDocs": "ts-node ./updateDocs.ts ",
-    "test": "jest",
-    "updateSnapshot": "jest --updateSnapshot"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "markdown-it": "^13.0.1",
-    "ts-node": "^10.9.1",
-    "yup": "^1.1.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.1",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.1.0"
-  }
-}
