+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
         "resolveJsonModule": true,
+        "esModuleInterop": true
     }
 }
